ai_learn/week_4/requirements.txt
sentence-transformers
faiss-cpu
numpy
tqdm
python-dotenv
openai